Made optuna module public without proxy module
